se termino la pantalla de carrito, faltan algunas cosas nada mas, como que ande el boton de seguircomprando, y el boton de carrito que te redireccione a la pantalla de carrito.

+++ ComprandoEnCasaBackEnd/src/main/java/com/ComprandoEnCasa/ComprandoEnCasaBackEnd/Repositories/ProductoRepository.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 @Repository
 public interface ProductoRepository extends CrudRepository<Producto, Integer> {
 
-    Optional<Producto> findById(Integer id);
+    Optional<Producto> findById(Long id);
 
     List<Producto> findAll();
